+++ lib/crs_cli/src/cmd__tools__emacs_grep.ml
@@ -76,7 +76,7 @@ let output_list crs ~oc ~repo_root ~below ~path_display_mode =
 
 let main =
   Command.make
-    ~summary:"A util used by the emacs crs-grep-mode."
+    ~summary:"A util for emacs crs-grep-mode to search CRs."
     ~readme:(fun () ->
       {|
 This command is a building block to the emacs $(b,crs-grep-mode).
@@ -106,8 +106,8 @@ By default the file paths are displayed relative to the command's $(b,cwd).
      let crs =
        let all_crs = Crs_parser.grep ~vcs ~repo_root ~below in
        match filters with
-       | `Default -> all_crs
-       | `Supplied filters ->
+       | `All -> all_crs
+       | `Only filters ->
          List.filter all_crs ~f:(fun cr ->
            List.exists filters ~f:(fun filter -> Cr_comment.Filter.matches filter ~cr))
      in